backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
websockets
google-generativeai
python-dotenv
# Visual Layer v2.5 dependencies
watchdog
httpx
python-socketio[asyncio]
aiofiles
# Database
psycopg2-binary
sqlalchemy

# Caching & Queue
redis
celery

# Monitoring
prometheus-fastapi-instrumentator

# AI & API
openai
google-generativeai

